refactor(tests): move the template cases to the boundaries module
Merging main plus this round's additions took the evidence suite back to
508 lines, over the 500-line file-size rule that an earlier review round
already caught once. The taste count ratchet flagged it again.

The cases moved are the ones that belong on the other side of the seam the
split already established: the canonical-template extraction, the
filled/unfilled pair, and the two drift guards all exercise text as pure
functions with no git repository and no staged index. 362 and 346 lines,
taste-lints 0 errors, 46 tests still passing.

The mutation harness already runs both files, so no mutant loses its
killer; only the recorded case count moves.

Refs #5205
